Mtp: fix NullReferenceException if playlistFolder wasn't found (bgo#733883)
There was already a way to try to prevent a NULL playlistFolder
in the MtpDevice class, but it was insufficient in my case. I'm
using a Nexus7, and in this case 'device.default_playlist_folder'
was not 0, but 4294967295 (which apparently is uint.MaxValue).

However, rather than adding yet another magic number in this if
statement, checking for a NULL playlistFolder at the end of the
foreach loop seems like a more robust approach.