Dap.Mtp: Improve status logging on load (bgo#734430)
The loading operation for MTP involves enumerating files, clearing
empty albums, writing track information to the database and reading
playlists. Banshee only produces a status update for the first step, in
addition the current never reaches the total (due to an off-by-one
issue) so it appears that the last track takes a long time to enumerate
and load.

This patch adds status messages for those operations so the user
doesn't experience a long delay with the same message before being
able to use his device. In addition, we move the calls to
Catalog.GetString outside of their respective loop bodies as a small
performance optimisation.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>