Podcasting: improve previous fix to not show the tweaked URL (bgo#729193)
The problem with previous fix was this: let's say the user tries
to add the URL for a podcast like this: ftp://mikrowelle.me/opus.xml.
As you can see, this URL already has a protocol, and what happens is
that the FTP access for that site is closed, so then banshee considers
it as not parseable, and with the previous fix, banshee would test
the URL "http://ftp://mikrowelle.me/opus.xml", and then it would
tell the user that there was a problem when parsing that last URL.

The way to fix this is to not overwrite the feedUri variable in the
second call to TryParseUrl(), so that, if an error has to be shown,
it will be about the URL that the user entered, not the tweaked one.