GStreamerSharp: fix audio&video playback in 32bits (bgo#750041)
The playback was broken on 32 bits because of a bug in gstreamer-sharp
where the handles always were 64 bits long even on 32 bits. This is
fixed in master [1] (which we already depend on, since [2]).

This fix updates VideoManager to account for the changes (and this
does not only fix video, since playing audio apparently also calls
WindowRealize).

[1] http://cgit.freedesktop.org/gstreamer/gstreamer-sharp/commit/?id=9556b05947a43cbeea57113f427d6bde5246c77f
[2] https://git.gnome.org/browse/banshee/commit/?id=e35092655886a6cce314d2310404f977233588a4

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>