GStreamerSharp: better window handle retrieval on OSX (bgo#749841)
On OSX, gdk can either be used with the quartz backend or the x11 one.
There is no function in gdk which returns which backend is currently used.
Therefore try to get the handle from the quartz backend first (which is more
common). A failure to find that function will result in an
EntryPointNotFoundException.

The bockbuild package always builds the quartz backend so this will always
works for a bockbuild build. The fallback is used when someone is still
using XQuartz and did compile all the dependencies and stuff on their own.

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>