osx: Update gtk-mac-integration to 2.0.7 (bgo#749839)
The current gtk-mac-integration needed for Banshee was
already 4 years old and doesn't compile anymore on osx 10.10.

The functions are now prefixed with gtkosx_ instead of gtk_osx.
From version 2.0.7 gtk-mac-integration is now split into two
libraries for gtk2 and gtk3.

Also update the configure check to test for the gtk3 version.

The relevant bockbuild commit that syncs with this change is:
https://github.com/BansheeMediaPlayer/bockbuild/commit/d1709f6193f65fa5fbbdc5bf92a9e148a8ffa700

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>