autogen.sh: use 'which' before running in run() function
A user just reported in the IRC channel that he had the error
'Could not run autoconf', even when he had it installed in
the system.

The problem is in the error message, because it's ambiguous,
it's not clear if 'autoconf' (in this case) was run but finished
with errors, or if it could not run because it was not found.

To fix this ambiguity, we change the error messages slightly
and check for the tool to run() with the command 'which' as a
sanity check.