Windows: Specify CallingConvention everywhere (bgo#751045)
The default CallingConvention is WinAPI, which defaults to
the system's convention which might either be StdCall or CDecl.
For C libraries the CallingConvention to use is CDecl
else strange runtime errors might occur because of stack
corruption.

Since Banshee is using C libraries only, specify the CDecl
CallinvConvention everywhere.

On Windows the MDA in Visual Studio would detect a wrong
calling convention (mono has no such detection).

(We take this opportunity to use const names for libraries
in some DllImport calls which point to the same lib.)

Signed-off-by: Andrés G. Aragoneses <knocte@gmail.com>