autogen.sh: call `make distclean` under the hood (bgo#741530)
As a sanity-check measure, it's useful that autogen.sh calls
`make distclean` in case there are binaries of a previous
compilation that could cause issues.

In particular, bgo#741530 is actually a consequence of this:
a developer (yours truly) switching to a stable branch to
run a different version of banshee, but forgetting to remove
the binaries from master-branch that were generated before,
therefore mixing assemblies from one branch and the other.

If this command fails because there were no makefiles at all
(which is what happens at a clean checkout) then it will of
course silently fail.