configure: depend on sqlite3 >= 3.8.1 (bgo#740879)
According to Roderich Schup (see [1]), the version check in
this commit from the master branch[2] was wrong.

Anyway, these days sqlite3 versions in distributions are
normally higher than 3.8.1 (e.g.: 3.11 in Ubuntu 16.04.1 LTS),
so we can raise the version dependency in the configure phase
to avoid the version check altogether, and just always use
UNLIKELY under the hood.

[1] https://bugzilla.gnome.org/show_bug.cgi?id=740879#c36
[2] "Database: fix performance for the new SQLite's query planner (bgo#740879)"
https://git.gnome.org/browse/banshee/commit/?id=fd3f08ec702f3a38918e0a96d2fc01d74af64ee4